--- a/src/@types/basic.d.ts
+++ b/src/@types/basic.d.ts
@@ -1,8 +1,6 @@
-import { Buffer } from 'buffer'
-
-export type bytes = Buffer
-export type bytes32 = Buffer
-export type bytes16 = Buffer
+export type bytes = Uint8Array
+export type bytes32 = Uint8Array
+export type bytes16 = Uint8Array
 
 export type uint32 = number
 export type uint64 = number