--- a/src/@types/handshake-interface.d.ts
+++ b/src/@types/handshake-interface.d.ts
@@ -5,7 +5,7 @@ import PeerId from 'peer-id'
 export interface IHandshake {
   session: NoiseSession
   remotePeer: PeerId
-  remoteEarlyData: Buffer
+  remoteEarlyData: bytes
   encrypt: (plaintext: bytes, session: NoiseSession) => bytes
   decrypt: (ciphertext: bytes, session: NoiseSession) => {plaintext: bytes, valid: boolean}
 }