--- a/src/@types/libp2p.d.ts
+++ b/src/@types/libp2p.d.ts
@@ -14,6 +14,6 @@ export interface INoiseConnection {
 
 export interface SecureOutbound {
   conn: any
-  remoteEarlyData: Buffer
+  remoteEarlyData: Uint8Array
   remotePeer: PeerId
 }