--- a/src/alloc-unsafe-browser.ts
+++ b/src/alloc-unsafe-browser.ts
@@ -1,0 +1,4 @@
+
+export default (size: number) => {
+  return new Uint8Array(size)
+}