--- a/src/alloc-unsafe-browser.ts
+++ b/src/alloc-unsafe-browser.ts
@@ -1,4 +1,0 @@
-
-export default (size: number) => {
-  return new Uint8Array(size)
-}