--- a/src/alloc-unsafe.ts
+++ b/src/alloc-unsafe.ts
@@ -1,0 +1,4 @@
+
+export default (size: number) => {
+  return Buffer.allocUnsafe(size)
+}