--- a/src/alloc-unsafe.ts
+++ b/src/alloc-unsafe.ts
@@ -1,4 +1,0 @@
-
-export default (size: number) => {
-  return Buffer.allocUnsafe(size)
-}