--- a/src/errors.ts
+++ b/src/errors.ts
@@ -1,9 +1,9 @@
 import BufferList from 'bl'
 
 export class FailedIKError extends Error {
-  public initialMsg: string|BufferList|Buffer
+  public initialMsg: string|BufferList|Uint8Array
 
-  constructor (initialMsg: string|BufferList|Buffer, message?: string) {
+  constructor (initialMsg: string|BufferList|Uint8Array, message?: string) {
     super(message)
 
     this.initialMsg = initialMsg