--- a/test/fixtures/node-globals.js
+++ b/test/fixtures/node-globals.js
@@ -1,2 +1,0 @@
-// @ts-nocheck
-export const { Buffer } = require('buffer')