--- a/test/utils.ts
+++ b/test/utils.ts
@@ -1,6 +1,7 @@
 import { keys, PrivateKey } from 'libp2p-crypto'
 import { KeyPair } from '../src/@types/libp2p'
 import PeerId from 'peer-id'
+import { Buffer } from 'buffer'
 
 export async function generateEd25519Keys (): Promise<PrivateKey> {
   return await keys.generateKeyPair('Ed25519', 32)