Replaces node buffers with browser-friendly `Uint8Array`s.

Previously we didn't want to do this because there's no equivalent to node's `Buffer.allocUnsafe` which doesn't set all entries of allocated buffers to 0.  The thing is we only actually use that in one place so to address this I've detected the existence of of `Buffer.allocUnsafe` before using it, otherwise falling back to `new Uint8Array` which is what the node `Buffer` polyfill does anyway.

Running the benchmark suite in this module shows the performance is comparable to or even slightly better than master (I think due to not having to convert `Uint8Array`s to `Buffer`s any more though please don't quote me on that 😆 ):

Before:

```console
$ node benchmark.js
Initializing handshake benchmark
Init complete, running benchmark
handshake x 59.95 ops/sec ±11.20% (75 runs sampled)
handshake x 54.68 ops/sec ±10.81% (68 runs sampled)
handshake x 50.42 ops/sec ±11.55% (65 runs sampled)
handshake x 53.41 ops/sec ±11.84% (68 runs sampled)
handshake x 50.25 ops/sec ±11.80% (66 runs sampled)
```

After:

```console
$ node ./benchmark.js
Initializing handshake benchmark
Init complete, running benchmark
handshake x 61.48 ops/sec ±11.71% (76 runs sampled)
handshake x 59.43 ops/sec ±11.13% (73 runs sampled)
handshake x 56.09 ops/sec ±12.02% (71 runs sampled)
handshake x 60.05 ops/sec ±11.69% (74 runs sampled)
handshake x 59.66 ops/sec ±10.59% (74 runs sampled)
```