@achingbrain 2022-01-03
That's interesting.  The percentage memory difference between `new Uint8Array(...)` and `Buffer.from(....)` seems to decrease as the size of the bytes increases so it may be some sort of fixed overhead, which is all weird as `Buffer` extends `Uint8Array`.  Maybe worth opening an issue against node itself to get some clarification?

As long as these things are getting garbage collected it shouldn't be a showstopper though?