@dapplion 2022-01-06
@achingbrain Sorry forgot to go back to this

> As long as these things are getting garbage collected it shouldn't be a showstopper though?

I did a few heap snapshots and I can't see noise retaining any substantial amount of binary data. So good on this front :+1: